fix the middleware pipeline
If a middleware is some thing like this:

````
router.Use(HandlerFunc(func(req *Request, next NextHandler) *response.Response {
    // a) do something before app logic
	resp := next(req)
    // b) do something after app logic
	return resp
}))
````

then we need to endure Render function should be called before #b. Otherwise the Render could fail if it depends on values in context. (for example, a) connect datasource and store it on context, then b) close connection